Saving ontologies in MappingSearch
